fast load np.int32

+++ lib/iris/tests/integration/fast_load/test_fast_load.py
@@ -377,7 +377,7 @@ class MixinCallDetails:
         # Make an 'expected' from selected fields, with the expected attribute.
         expected = CubeList([flds[1], flds[3]]).merge()
         if not self.do_fast_loads:
-            expected[0].attributes["LBVC"] = 8
+            expected[0].attributes["LBVC"] = np.int32(8)
         else:
             expected[0].attributes["A_LBVC"] = [8, 8]
